+++ comments.php
@@ -14,24 +14,30 @@ if ( post_password_required() || ( !have_comments() && !comments_open() && !ping
 			<h4 class="comments-title">Reacties</h4>
 
 			<ol class="comment-list">
-				<?php wp_list_comments(); ?>
+				<?php wp_list_comments(
+					array(
+						'callback'     => 'hybrid_comments_callback',
+						'end-callback' => 'hybrid_comments_end_callback'
+					)
+				); ?>
 			</ol><!-- .comment-list -->
 
-			<?php locate_template( array( 'partials/comments-nav.php' ), true ); // Loads the partials/comments-nav.php template. ?>
+			<?php locate_template( array( 'partials/comment/comments-nav.php' ), true ); // Loads the partials/comment/comments-nav.php template. ?>
 
 		</div><!-- #comments-->
 
 	<?php endif; // End check for comments. ?>
 
-	<?php locate_template( array( 'partials/comments-error.php' ), true ); // Loads the partials/comments-error.php template. ?>
+	<?php locate_template( array( 'partials/comment/comments-error.php' ), true ); // Loads the partials/comment/comments-error.php template. ?>
 
 	<?php
 		$args = array(
-			'cancel_reply_link' => '<i class="fa fa-remove"></i>',
+			'cancel_reply_link' => '<span class="icon icon-remove"></span>',
+			'cancel_reply_before' => '',
+			'cancel_reply_after' => '',
 			'comment_notes_before' => '',
 			'comment_notes_after' => '',
 		);
-
 	?>
 
 	<?php comment_form( $args ); // Loads the comment form. ?>